+++ templates/hexarch/API/API.Adapters.Test/Fixture/ApplicationFixture.cs
@@ -4,12 +4,14 @@ using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Data;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Db;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.DependencyInjection;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Test.Fixtures;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Ports;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Hosting;
 
 namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.API.Adapters.Test.Fixture;
 
@@ -25,6 +27,8 @@ public class ApplicationFixture : WebApplicationFactory<IAssemblyInfo>, IAsyncLi
     {
         builder.ConfigureTestServices(services =>
         {
+            // remove background jobs from startup
+            services.RemoveAll<IHostedService>();
             services.RemoveAll<InfrastructureConfiguration>();
             services.AddSingleton(
                 new InfrastructureConfiguration
@@ -41,6 +45,14 @@ public class ApplicationFixture : WebApplicationFactory<IAssemblyInfo>, IAsyncLi
         });
     }
 
+    private async Task RunMigrations()
+    {
+        using var scope = Services.CreateScope();
+        var migrationService =
+            scope.ServiceProvider.GetRequiredService<IDatabaseMigrationService>();
+        await migrationService.RunMigrations();
+    }
+
     private async Task SeedDatabase()
     {
         using var scope = Services.CreateScope();
@@ -51,13 +63,14 @@ public class ApplicationFixture : WebApplicationFactory<IAssemblyInfo>, IAsyncLi
         );
     }
 
-    public async Task InitializeAsync()
+    async ValueTask IAsyncLifetime.InitializeAsync()
     {
         await _postgresDbDemoFixture.InitializeAsync();
+        await RunMigrations();
         await SeedDatabase();
     }
 
-    public new async Task DisposeAsync()
+    async ValueTask IAsyncDisposable.DisposeAsync()
     {
         await _postgresDbDemoFixture.DisposeAsync();
     }