+++ templates/hexarch/API/API.Adapters.Test/Fixture/ApplicationFixtureCollection.cs
@@ -1,0 +1,6 @@
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.API.Adapters.Test.Fixture;
+
+namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.API.Adapters.Test;
+
+[CollectionDefinition("Application Fixture Collection")]
+public class ApplicationFixtureCollection : ICollectionFixture<ApplicationFixture> { }